requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-dotenv==1.0.1
openai==1.40.6
azure-identity==1.17.1
langdetect==1.0.9
tiktoken==0.7.0
gunicorn==21.2.0
python-docx==1.1.2
PyPDF2==3.0.1
